Use the AES implementation that comes with I2P instead of a JCE provider because the Sun JRE makes it a PITA to enable "unlimited strength encryption" in an end user friendly manner. Fortunately, only AES is affected by the crypto restrictions.
Another problem with the Sun JRE is that it makes it near impossible to ship only a subset of the 1.7 meg BouncyCastle .jar. The .jar has to be signed with a key certified by Sun, which is a lengthy process that involves giving Sun a DNA sample and sacrificing an animal in front of their headquarters. So the ECC code now uses the BouncyCastle API instead of JCE.